+++ src/test/java/com/zepben/evolve/cim/iec61970/base/wires/generation/production/BatteryUnitTest.kt
@@ -31,8 +31,8 @@ internal class BatteryUnitTest {
         val batteryUnit = BatteryUnit()
 
         MatcherAssert.assertThat(batteryUnit.batteryState, Matchers.nullValue())
-        MatcherAssert.assertThat(batteryUnit.ratedE, Matchers.equalTo(0))
-        MatcherAssert.assertThat(batteryUnit.storedE, Matchers.equalTo(0))
+        MatcherAssert.assertThat(batteryUnit.ratedE, Matchers.equalTo(0.0))
+        MatcherAssert.assertThat(batteryUnit.storedE, Matchers.equalTo(0.0))
 
         batteryUnit.apply {
             this.batteryState = BatteryStateKind.charging